Create integrationEventTopicService
